remove deprecated and unused variables improve the organization code quality
